+++ nens_auth_client/oauth.py
@@ -1,0 +1,4 @@
+from authlib.integrations.django_client import OAuth
+
+# Create the global OAuth registry
+oauth = OAuth()