+++ nens_auth_client/urls.py
@@ -16,10 +16,15 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path
+from django.conf import settings
 
 from nens_auth_client import views
 
-
 urlpatterns = [
-    path('admin/', admin.site.urls),
+    path("authorize", views.authorize, name="authorize"),
+    path("login", views.login, name="login"),
+    path("logout", views.logout, name="logout"),
 ]
+
+if settings.NENS_AUTH_STANDALONE:
+    urlpatterns += [path("admin/", admin.site.urls)]