+++ CHANGES.rst
@@ -7,3 +7,8 @@ Changelog of nens-auth-client
 
 - Initial project structure created with cookiecutter and
   https://github.com/nens/cookiecutter-djangosite-template
+
+- Added SocialUser model.
+
+- Added login, authorize and logout views for login/logout via OpenID Connect
+  (AWS Cognito). Default settings are setup via django-appconf.